Add files via upload
In the first course on Linear Algebra, we look at what linear algebra is and how it relates to data. Then we look through what vectors and matrices are and how to work with them.

The second course, Multivariate Calculus, builds on this to look at how to optimize fitting functions to get good fits to data. It starts from introductory calculus and then uses the matrices and vectors from the first course to look at data fitting, wonderfully explained the neural networks.

The third course, Dimensionality Reduction with Principal Component Analysis, uses the mathematics from the first two courses to compress high-dimensional data. This course is of intermediate difficulty and will require Python and NumPy knowledge. Little bit complex as it has a combination of both first course and the second course, graphically explained inner products, orthogonal projection.

The course could be certainly beneficial if we go through
3Blue1Brown Essence of linear algebra, Essence of calculus.
https://lnkd.in/df-aiyA

https://lnkd.in/ddVpVeH